requirements.txt
python-gitlab
bottle
Adafruit_Python_CharLCD
